исправил работу с exceptions в auth

+++ src/api/auth.py
@@ -1,62 +1,46 @@
 
-from fastapi import APIRouter, HTTPException, Response
+from fastapi import APIRouter, Response, Request
 
-from src.exceptions import UserWithSuchEmailAlreadyExists, ObjectAlreadyExistsException
-from src.Schemas.users import UserRequestAdd, UserAdd
+from src.exceptions import UserAlreadyExistsException, UserWithSuchEmailAlreadyExistsHTTPExceptions, \
+    UserNotFoundException, UserNotFoundHTTPException, UserAlreadyLogInHTTPException, UserAlreadyLogInException, \
+    UserAlreadyLogOutHTTPException
+from src.Schemas.users import UserRequestAdd
 from src.api.dependencies import UserIdDep, DBDep
 
 from src.services.auth import AuthService
-
-
-
 
 router = APIRouter(prefix="/auth", tags=["Авторизация и аутентификация"])
 
 @router.post("/register")
 async def register_user(data: UserRequestAdd, db:DBDep):
-
-
-    hashed_password = AuthService().hash_password(data.password)
-    new_user_data = UserAdd(email= data.email, hashed_password = hashed_password)
     try:
-        await db.users.add(new_user_data)
-        await db.commit()
-    except ObjectAlreadyExistsException:
-        raise UserWithSuchEmailAlreadyExists
-
-
-
-    return {"Status": "OK"}
-
-
-
-
-
+        await AuthService(db).register_user(data)
+    except UserAlreadyExistsException:
+        raise UserWithSuchEmailAlreadyExistsHTTPExceptions
+    return {f"Пользователь: {data.email} успешно зарегистрирован"}
 
 @router.post("/login")
-async def login_user(data: UserRequestAdd, response:Response, db:DBDep):
+async def login_user(data: UserRequestAdd, response:Response, db:DBDep, request:Request):
+    try:
+        request.cookies.get("access_token")
+        access_token = await AuthService(db).login_user(data)
+    except UserNotFoundException:
+        raise UserNotFoundHTTPException
+    except UserAlreadyLogInException:
+        raise UserAlreadyLogInHTTPException
 
-
-    user = await db.users.get_user_with_hashed_password(email = data.email)
-    if not user:
-        raise HTTPException(status_code = 401, detail = "Пользователь с таким email не существует")
-    if not AuthService().verify_password(data.password, user.hashed_password):
-        raise HTTPException(status_code=401, detail="Пароль неверный")
-
-    access_token = AuthService().create_access_token({"user_id": user.id})
     response.set_cookie("access_token", access_token)
-    return {"access_token": access_token}
+    return {"Вы вошли как ":data.email, "access_token": access_token}
 
 
 @router.get("/me")
 async def me(user_id:UserIdDep, db:DBDep):
-
-
-    user = await db.users.get_one_or_none(id = user_id)
+    user = await AuthService(db).get_me(user_id)
     return user
 
 @router.post("/logout")
-async def logout_user(response:Response):
-
+async def logout_user(response:Response, request:Request):
+    if not request.cookies.get("access_token"):
+        raise UserAlreadyLogOutHTTPException
     response.delete_cookie("access_token")
     return {"Status":"OK"}
